+++ src/api/types/rest.ts
@@ -35,6 +35,8 @@ export interface Query extends StrictQuery {
 
 export interface QueryObject extends Query {
     data?: "Disallow deprecated api.Sample.queryOne({ data: data.id }). Use api.Data.getSampleFromDataId(data.id) instead.";
+    parents?: "Disallow deprecated api.Data.query({ parents: data.id }). Use api.Data.getChildren(data.id) instead.";
+    children?: "Disallow deprecated api.Data.query({ children: data.id }). Use api.Data.getParents(data.id) instead.";
     hydrate_data?: void;
 }